requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
backports.zoneinfo==0.2.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Deprecated==1.2.13
Django==4.1.2
django-appconf==1.0.5
django-celery-email==3.0.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-extra-fields==3.0.2
django-phonenumber-field==7.0.0
django-phonenumbers==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
eventlet==0.33.2
greenlet==2.0.1
idna==3.4
kombu==5.2.4
packaging==21.3
phonenumbers==8.13.0
Pillow==9.3.0
prompt-toolkit==3.0.32
psycopg2==2.9.5
pycryptodome==3.15.0
PyJWT==2.6.0
pyparsing==3.0.9
python-decouple==3.6
pytz==2022.5
redis==4.3.4
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
